@dinamic 2019-01-24
Wow, that's good 100% I/O improvement! I think it would have big impact on the performance of software that relies on xhyve - take docker for macOS as an example.